Posodobil Kodirnik/Dekodirnik za UI/Demo da sta ista kot za DLL. Odstranil neuporabljen RangeComarer, ker sem implementiral IComparable kar v Simbolu. Popravil desno shiftanje pri ULong (UInt64), da ni problemov če shiftamo za več kot 32 bitov.

+++ BinIO/BitWriter.cs
@@ -48,7 +48,7 @@ namespace BinIO {
             //zapišemo v tren bajt
             else{
                 //preberemo "bits" bitov
-                ulong maska = (ulong) ((1 << bits) - 1);
+                ulong maska = (1UL << bits) - 1;
 
                 //premaknemo že vpisano na levo
                 _currByte <<= bits;
@@ -62,7 +62,7 @@ namespace BinIO {
 
         private void FillByte(ulong data, byte bitsToFill) {
             //pridobimo toliko "bitsToFill" bitov iz "data"
-            ulong maska = (ulong) ((1 << bitsToFill) - 1);
+            ulong maska = (1UL << bitsToFill) - 1;
 
             //potisnemo že zapisano na levo
             _currByte <<= bitsToFill;
